+++ packages/react-front-kit-shared/src/types/filters.ts
@@ -1,0 +1,8 @@
+import type { ReactElement } from 'react';
+
+export interface IFilter {
+  active?: boolean;
+  component?: ReactElement;
+  id: number | string;
+  label: string;
+}