+++ packages/react-front-kit-shared/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './actions';
 export * from './options';
 export * from './theme';
+export * from './filters';