+++ packages/react-front-kit/src/Components/SidebarFilters/SidebarFilters.test.tsx
@@ -1,13 +1,13 @@
 import { renderWithProviders } from '@smile/react-front-kit-shared/test-utils';
 
-import { FiltersBar } from './FiltersBar';
-import { activeFilters } from './FitlersBar.mock';
 import { menuWithoutContent } from './SidebarFilterMenu/SidebarFilterMenu.mock';
+import { SidebarFilters } from './SidebarFilters';
+import { activeFilters } from './SidebarFilters.mock';
 
 describe('FiltersBar', () => {
   it('matches snapshot', () => {
     const { container } = renderWithProviders(
-      <FiltersBar
+      <SidebarFilters
         activeFilters={activeFilters}
         defaultOpenedMenuIds={[1, 3, 10, 11]}
         deleteButtonLabel="Remove all"